cqrs-python-debezium-opensearch/projector-service/requirements.txt
kafka-python==2.3.1
opensearch-py==2.6.0
